Adjusted inflation calculations
Added array for average monthly housing prices to central bank agent.
Changed the order around for sending the housing prices to the central
bank and the central bank calculating the inflation and sending it back
to the households for calculating the inflated principle. Changed how
inflation is calculated, adding the weighted price changes together
instead of the prices.